+++ eth_defi/maseer_one/__init__.py
@@ -1,0 +1,1 @@
+"""Maseer One tokenised asset support."""